requirements.txt
GitPython
loguru
pandas
pyfiglet
requests
termcolor